services/backend/requirements.txt
fastapi==0.128.0
fastapi-limiter==0.1.6
uvicorn[standard]==0.40.0
gunicorn==23.0.0
redis==7.1.0
sqlalchemy==2.0.45
asyncpg==0.31.0
alembic==1.17.2
pydantic-settings==2.12.0
python-dotenv==1.2.1
psycopg2-binary==2.9.11
pytest==9.0.2
httpx==0.28.1
python-jose[cryptography]==3.5.0
passlib[bcrypt,argon2]==1.7.4
argon2-cffi==25.1.0
python-multipart==0.0.21
email-validator==2.3.0